Create organization form component
