+++ src/Components/Organizations/OrganizationForm.css
@@ -1,0 +1,5 @@
+.organization-form__actions {
+  display: flex;
+  gap: 1rem;
+  justify-content: flex-end;
+}
