+++ src/Components/Organizations/OrganizationForm.jsx
@@ -1,0 +1,33 @@
+import React, { useState } from "react";
+import TextBox from "..//design/TextBox/TextBox";
+import Button from "../design/Button/Button";
+import "./OrganizationForm.css";
+
+export default function OrganizationForm(props) {
+  const { onSubmit, onCancel } = props;
+  const [name, setName] = useState("");
+
+  const handleSubmit = async (event) => {
+    event.preventDefault();
+    onSubmit({ name });
+  };
+
+  return (
+    <form className="organization-form" onSubmit={handleSubmit}>
+      <TextBox
+        labelText="Organization Name"
+        type="text"
+        name="name"
+        value={name}
+        onChange={(event) => setName(event.target.value)}
+        required
+      />
+      <div className="organization-form__actions">
+        <Button variant="text" onClick={onCancel}>
+          Cancel
+        </Button>
+        <Button type="submit">Add</Button>
+      </div>
+    </form>
+  );
+}
